requirements.txt
certifi==2018.8.13
chardet==3.0.4
idna==2.7
mongoengine==0.15.3
oauthlib==2.1.0
pymongo==3.7.1
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
six==1.11.0
twitter==1.18.0
urllib3==1.23
